% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{EXAMPLE_PYTHON_DIR}
\alias{EXAMPLE_PYTHON_DIR}
\title{Package directory containing example python scripts}
\format{
An object of class \code{character} of length 1.
}
\usage{
EXAMPLE_PYTHON_DIR
}
\description{
Package directory containing example python scripts
}
\keyword{internal}
