% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-python-pkgs.R
\name{install_py_pkgs}
\alias{install_py_pkgs}
\title{Install python packages to an environment}
\usage{
install_py_pkgs(
  py_pkgs = c("scipy", "pandas"),
  py_env = c("conda", "virtual"),
  env_name = PYTHON_R_ENV,
  update = FALSE,
  python_version = NULL
)
}
\arguments{
\item{py_pkgs}{vector of python packages to install}

\item{py_env}{Environment type. Either a conda or virtual environment.}

\item{env_name}{an environment name to use for \strong{new} virtual or conda environments.}

\item{update}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, update packages that are
already installed. Otherwise skip their installation.}

\item{python_version}{The requested Python version. Ignored when attempting
to install with a Python virtual environment.}
}
\description{
Install python packages to an environment
}
\details{
see \code{?reticulate::py_install} for more details
}
\seealso{
\link{setup_py_env}
}
