% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{is_miniconda_installed}
\alias{is_miniconda_installed}
\title{Check that minicoonda is installed at specified location}
\usage{
is_miniconda_installed(path = reticulate::miniconda_path())
}
\arguments{
\item{path}{path to miniconda installation}
}
\description{
Check that minicoonda is installed at specified location
}
\note{
the assumed extension changes depending on operating system.
}
\keyword{internal}
