% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{miniconda_is_installed}
\alias{miniconda_is_installed}
\title{Check that minicoonda is installed at specified location}
\usage{
miniconda_is_installed()
}
\arguments{
\item{path}{path to miniconda installation}
}
\description{
Check that minicoonda is installed at specified location
}
\note{
the assumed extension changes depending on operating system.
}
\keyword{internal}
