% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{py_add}
\alias{py_add}
\title{Simple python function}
\format{
An object of class \code{character} of length 1.
}
\usage{
py_add
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\description{
Simple python function
}
\details{
This function does not require imports or virtual environment (much faster)
}
\keyword{datasets}
