% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{py_array}
\alias{py_array}
\title{Create list of arrays using pandas}
\format{
An object of class \code{character} of length 1.
}
\usage{
py_array
}
\description{
Create list of arrays using pandas
}
\keyword{datasets}
