% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{py_check_path}
\alias{py_check_path}
\title{Check that file exists at location using python}
\format{
An object of class \code{character} of length 1.
}
\usage{
py_check_path
}
\arguments{
\item{file_path}{path to audio file}
}
\description{
Example with working with \code{file_path} and importing a python package that
\emph{comes with python}
}
\details{
The \code{os} python module comes with python:
\itemize{
\item Virtual environment (or conda) is only needed when we have to \emph{install}
packages.
}
}
\keyword{internal}
