% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-python-pkgs.R
\name{py_module_available}
\alias{py_module_available}
\title{Modified version of \code{\link[reticulate:py_module_available]{reticulate::py_module_available()}}}
\usage{
py_module_available(module, path)
}
\arguments{
\item{module}{The name of the Python module.}

\item{path}{The path from which the module should be imported.}
}
\description{
Modified version of \code{\link[reticulate:py_module_available]{reticulate::py_module_available()}}
}
\keyword{internal}
