% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{python_is_installed}
\alias{python_is_installed}
\title{Check that python is installed}
\usage{
python_is_installed(use_environment = NULL)
}
\arguments{
\item{use_environment}{Logical (\code{TRUE}/\code{FALSE}). Passed to
\code{\link[reticulate:py_discover_config]{reticulate::py_discover_config()}}}
}
\description{
Check that python is installed
}
