% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-python-pkgs.R
\name{check_py_pkgs_installed}
\alias{check_py_pkgs_installed}
\title{Helper function for ensuring required python packages are installed}
\usage{
check_py_pkgs_installed(py_pkgs)
}
\arguments{
\item{py_pkgs}{vector of python packages to check that are installed.}
}
\description{
Helper function for ensuring required python packages are installed
}
