% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{get_py_example}
\alias{get_py_example}
\alias{pythonR_examples}
\title{Retrieve example function from \code{pythonR_examples}}
\format{
An object of class \code{environment} of length 3.
}
\usage{
get_py_example(example = "py_add")

pythonR_examples
}
\arguments{
\item{example}{The name of an example to load. See \code{ls(pythonR_examples)} for
options.}
}
\description{
Retrieve example function from \code{pythonR_examples}

Environment containing \code{pythonR} example functions
}
\examples{
\dontrun{
# Get list of available examples.
ls(pythonR_examples)


# View the functions
get("py_add", envir = pythonR_examples)
get("py_check_path", envir = pythonR_examples)
get("py_array", envir = pythonR_examples)

# Assign the function to use it
py_add <- get_py_example("py_add")
py_add(3, 4)
}

}
\keyword{datasets}
